Moved ride create command to new game action system.
