+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -1,6 +1,7 @@
 #include "GameAction.h"
 #include "PlaceParkEntranceAction.hpp"
 #include "SetParkEntranceFeeAction.hpp"
+#include "RideCreateAction.hpp"
 
 namespace GameActions {
 
@@ -8,6 +9,7 @@ void Register()
 {
     Register<SetParkEntranceFeeAction>();
     Register<PlaceParkEntranceAction>();
+    Register<RideCreateAction>();
 }
 
 }
